feat: #493 Inconsistent abstractions / Constants & config fix
