+++ apps/web/components/Feature/Dashboard/ClientDashboardViewer.tsx
@@ -177,8 +177,8 @@ export default function ClientDashboardViewer() {
         onConfirm={handleConfirmLogout}
         onCancel={handleCancelLogout}
         onClose={handleCancelLogout}
-        width="480px"
-        padding="40px 32px"
+        size="sm"
+        spacing="sm"
         buttonAlign="center"
         textAlign="center"
         showCloseButton={false}
