+++ apps/web/components/Feature/Dashboard/ClientViewerProfile/index.tsx
@@ -124,7 +124,7 @@ export default function ClientViewerProfile() {
         onClose={handlePasswordClose}
         onCancel={handlePasswordClose}
         onConfirm={handlePasswordSave}
-        width="630px"
+        size="md"
         fullWidthButtons
         buttonRow
         confirmDisabled={!isPasswordFormValid}
@@ -149,7 +149,7 @@ export default function ClientViewerProfile() {
         confirmLabel={t("nav.login")}
         onClose={() => setShowPasswordSuccessModal(false)}
         onConfirm={() => router.push(PATHS.AUTH_LOGIN)}
-        width="480px"
+        size="sm"
         showCloseButton={false}
       />
     </Container>
