+++ apps/web/utils/keyboard.ts
@@ -1,0 +1,9 @@
+const KEY_SPACE = " ";
+const KEY_SPACEBAR = "Spacebar";
+
+export const KEY_ENTER = "Enter";
+export const ESCAPE = "Escape";
+export const KEYDOWN = "keydown";
+export const BUTTON = "button";
+export const POINTER_DOWN = "pointerdown";
+export const FAQ_TOGGLE_KEYS = [KEY_ENTER, KEY_SPACE, KEY_SPACEBAR];
